requirements.txt
django==3.2.13
django-widget-tweaks==1.4.12
django-crispy-forms==1.14.0
python-decouple==3.6
psycopg2-binary
django-auditlog==2.0.0
djangorestframework==3.13.1
crispy_bootstrap4
weasyprint==55.0

